Bug 1667032 - Fix signed-vs-unsigned comparison in nsPrinterWin::CopyDefaultDevmodeW(). r=emilio
Differential Revision: https://phabricator.services.mozilla.com/D91256

UltraBlame original commit: 1ecd9c203a7c6c7dfc56b38fee617c1e1d0c165a

+++ widget/windows/nsPrinterWin.cpp
@@ -46,9 +46,11 @@ nsTArray<uint8_t> nsPrinterWin::CopyDefaultDevmodeW() const {
     
     LONG bytesNeeded = ::DocumentPropertiesW(nullptr, autoPrinter.get(),
                                              mName.get(), nullptr, nullptr, 0);
-    MOZ_DIAGNOSTIC_ASSERT(bytesNeeded >= sizeof(DEVMODEW),
+    
+    
+    MOZ_DIAGNOSTIC_ASSERT(bytesNeeded >= LONG(sizeof(DEVMODEW)),
                           "DocumentPropertiesW failed to get valid size");
-    if (bytesNeeded < sizeof(DEVMODEW)) {
+    if (bytesNeeded < LONG(sizeof(DEVMODEW))) {
       return devmodeStorageW;
     }
